move schema back because docs depend on it
