+++ python/ray/serve/_private/client.py
@@ -20,7 +20,7 @@ from ray.serve.exceptions import RayServeException
 from ray.serve.generated.serve_pb2 import DeploymentRoute, DeploymentRouteList
 from ray.serve.generated.serve_pb2 import StatusOverview as StatusOverviewProto
 from ray.serve.handle import RayServeHandle, RayServeSyncHandle
-from ray.serve._private.schema import ServeApplicationSchema
+from ray.serve.schema import ServeApplicationSchema
 
 logger = logging.getLogger(__file__)
 # Whether to issue warnings about using sync handles in async context
